+++ DIYproject/src/DIYDriver.java
@@ -1,0 +1,9 @@
+public class DIYDriver {
+
+	@SuppressWarnings("deprecation")
+	public static void main(String[] args) {
+		DIYMain control = new DIYMain();	
+		control.show();
+	}
+
+}
