+++ DIYproject/src/DIYMain.java
@@ -1,9 +1,32 @@
-public class DIYMain {
+import java.awt.FlowLayout;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import javax.swing.*;
 
-	@SuppressWarnings("deprecation")
-	public static void main(String[] args) {
-		DIYController control = new DIYController();	
-		control.show();
+public class DIYMain extends JFrame {
+
+	private static final long serialVersionUID = 1L;
+	private JButton myButton; 
+	private JPanel myMainPanel;
+	
+	public DIYMain() {
+		super();
+		setSize(400, 800);
+		setDefaultCloseOperation( JFrame.EXIT_ON_CLOSE );
+		setLayout(new FlowLayout());
+		myMainPanel = new JPanel();
+		myButton = new JButton("About");
+		myButton.addActionListener(new ActionListener() {
+			@Override
+			public void actionPerformed(final ActionEvent event) {
+				DIYAbout about = new DIYAbout();
+				about.showAbout();
+			}
+		});
+		add(myMainPanel);
+		add(myButton);
+		setVisible(true);
+		pack();
 	}
-
+	
 }
